pyproject.toml
[project]
name = "t212-client-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.3",
]

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
extend-select = ["E501", "N", "R", "I", "C901", "PD", "NPY"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.mypy]
strict = true
explicit_package_bases = true
check_untyped_defs = true
disable_error_code = ["no-any-return"]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.0",
]
